Genrate Report
Las estadísticas dan cualquier cosa, solamente está hecha la estructura para mostrar los resultados

+++ src/main/java/entities/Entity.java
@@ -14,8 +14,8 @@ public abstract class Entity {
 
     // attributes
     private int id;
-    private int waitingTime;
-    private int transitTime;
+    private double waitingTime;
+    private double transitTime;
 
     // associations
     private Server attendingServer;
@@ -50,19 +50,19 @@ public abstract class Entity {
         return this.id;
     }
 
-    public int getWaitingTime() {
+    public double getWaitingTime() {
         return waitingTime;
     }
 
-    public void setWaitingTime(int waitingTime) {
+    public void setWaitingTime(double waitingTime) {
         this.waitingTime = waitingTime;
     }
 
-    public int getTransitTime() {
+    public double getTransitTime() {
         return this.transitTime;
     }
 
-    public void setTransitTime(int transitTime) {
+    public void setTransitTime(double transitTime) {
         this.transitTime = transitTime;
     }
